+++ apps/whispering/src/lib/services/db/index.ts
@@ -1,13 +1,13 @@
-import { DownloadServiceLive } from '../download';
-import { createDbServiceDexie } from './dexie';
+import { DownloadServiceLive } from '$lib/services/download';
+import { createDbServiceDexie } from '$lib/services/db/dexie';
 
-export { createDbServiceDexie, DbServiceErr } from './dexie';
+export { createDbServiceDexie, DbServiceErr } from '$lib/services/db/dexie';
 export {
 	generateDefaultTransformation,
 	generateDefaultTransformationStep,
 	TRANSFORMATION_STEP_TYPES,
 	TRANSFORMATION_STEP_TYPES_TO_LABELS,
-} from './models';
+} from '$lib/services/db/models';
 export type {
 	InsertTransformationStep,
 	Recording,
@@ -17,7 +17,7 @@ export type {
 	TransformationRunFailed,
 	TransformationStep,
 	TransformationStepRun,
-} from './models';
+} from '$lib/services/db/models';
 
 export const DbServiceLive = createDbServiceDexie({
 	DownloadService: DownloadServiceLive,
