+++ apps/whispering/src/lib/services/ffmpeg/web.ts
@@ -1,6 +1,6 @@
 import { Ok, type Result } from 'wellcrafted/result';
 import type { WhisperingError } from '$lib/result';
-import type { FfmpegService } from './types';
+import type { FfmpegService } from '$lib/services/ffmpeg/types';
 
 export function createFfmpegServiceWeb(): FfmpegService {
 	return {
