+++ apps/whispering/src/lib/services/text/index.ts
@@ -1,7 +1,7 @@
-import { createTextServiceDesktop } from './desktop';
-import { createTextServiceWeb } from './web';
+import { createTextServiceDesktop } from '$lib/services/text/desktop';
+import { createTextServiceWeb } from '$lib/services/text/web';
 
-export type { TextService, TextServiceError } from './types';
+export type { TextService, TextServiceError } from '$lib/services/text/types';
 
 export const TextServiceLive = window.__TAURI_INTERNALS__
 	? createTextServiceDesktop()
