+++ apps/whispering/src/lib/utils/check-ffmpeg.ts
