+++ apps/whispering/src/lib/utils/register-commands.ts
